Updated PR by reason of that JUnit executes method evaluate() after method before() and it means that all pre-conditions will not be skipped even if the test is skipped.
