+++ common-test/src/main/java/io/strimzi/test/JUnitGroup.java
@@ -13,6 +13,13 @@ import java.lang.annotation.Target;
  * Annotation is intended for test methods for the ability to group test cases and execute certain types of groups on CI.
  * For using this annotation, need to define an instance of ({@link JUnitGroupRule}) as a ({@link org.junit.Rule})
  * for each test classes.
+ *
+ * Availible types - acceptance, regression
+ *
+ * How to execute the expected group of system tests. Add system property "junitgroup" with following value:
+ * -Djunitgroup=integration - to execute one test group
+ * -Djunitgroup=acceptance,regression - to execute many test groups
+ * -Djunitgroup=all - to execute all test groups
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.METHOD, ElementType.TYPE})
@@ -28,5 +35,5 @@ public @interface JUnitGroup {
      * Name of the test group(s) to which a test belongs. If the name is not defined, the test will belong to the default
      * implicit group.
      */
-    String[] value() default {};
+    String[] name() default {};
 }
