+++ systemtest/src/test/java/io/strimzi/systemtest/AbstractClusterIT.java
@@ -14,14 +14,12 @@ import io.fabric8.kubernetes.api.model.Pod;
 import io.fabric8.kubernetes.api.model.Quantity;
 import io.fabric8.kubernetes.client.DefaultKubernetesClient;
 import io.fabric8.kubernetes.client.KubernetesClient;
-import io.strimzi.test.JUnitGroupRule;
 import io.strimzi.test.TestUtils;
 import io.strimzi.test.k8s.KubeClient;
 import io.strimzi.test.k8s.KubeClusterException;
 import io.strimzi.test.k8s.KubeClusterResource;
 import io.strimzi.test.k8s.ProcessResult;
 import org.junit.ClassRule;
-import org.junit.Rule;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -48,9 +46,6 @@ public class AbstractClusterIT {
 
     @ClassRule
     public static KubeClusterResource cluster = new KubeClusterResource();
-
-    @Rule
-    public JUnitGroupRule rule = new JUnitGroupRule();
 
     static KubernetesClient client = new DefaultKubernetesClient();
     KubeClient<?> kubeClient = cluster.client();
