+++ systemtest/src/test/java/io/strimzi/systemtest/OpenShiftTemplatesIT.java
@@ -9,7 +9,6 @@ import io.fabric8.kubernetes.api.model.ConfigMap;
 import io.fabric8.kubernetes.client.DefaultKubernetesClient;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.strimzi.test.JUnitGroup;
-import io.strimzi.test.JUnitGroupRule;
 import io.strimzi.test.Namespace;
 import io.strimzi.test.OpenShiftOnly;
 import io.strimzi.test.Resources;
@@ -17,7 +16,6 @@ import io.strimzi.test.StrimziRunner;
 import io.strimzi.test.k8s.KubeClusterResource;
 import io.strimzi.test.k8s.Oc;
 import org.junit.ClassRule;
-import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
@@ -45,15 +43,12 @@ public class OpenShiftTemplatesIT {
     @ClassRule
     public static KubeClusterResource cluster = new KubeClusterResource();
 
-    @Rule
-    public JUnitGroupRule rule = new JUnitGroupRule();
-
     private ObjectMapper mapper = new ObjectMapper();
     private Oc oc = (Oc) cluster.client();
     private KubernetesClient client = new DefaultKubernetesClient();
 
     @Test
-    @JUnitGroup(value = {"acceptance"})
+    @JUnitGroup(name = "acceptance")
     public void testStrimziEphemeral() throws IOException {
         String clusterName = "foo";
         oc.newApp("strimzi-ephemeral", map("CLUSTER_NAME", clusterName,
@@ -70,7 +65,7 @@ public class OpenShiftTemplatesIT {
     }
 
     @Test
-    @JUnitGroup(value = {"acceptance"})
+    @JUnitGroup(name = "acceptance")
     public void testStrimziPersistent() throws IOException {
         String clusterName = "bar";
         oc.newApp("strimzi-persistent", map("CLUSTER_NAME", clusterName,
@@ -87,7 +82,7 @@ public class OpenShiftTemplatesIT {
     }
 
     @Test
-    @JUnitGroup(value = {"acceptance"})
+    @JUnitGroup(name = "acceptance")
     public void testStrimziEphemeralWithCustomParameters() {
         String clusterName = "test-ephemeral-with-custom-parameters";
         oc.newApp("strimzi-ephemeral", map("CLUSTER_NAME", clusterName,
@@ -113,7 +108,7 @@ public class OpenShiftTemplatesIT {
     }
 
     @Test
-    @JUnitGroup(value = {"acceptance"})
+    @JUnitGroup(name = "acceptance")
     public void testStrimziPersistentWithCustomParameters() throws IOException {
         String clusterName = "test-persistent-with-custom-parameters";
         oc.newApp("strimzi-persistent", map("CLUSTER_NAME", clusterName,
@@ -143,7 +138,7 @@ public class OpenShiftTemplatesIT {
     }
 
     @Test
-    @JUnitGroup(value = {"acceptance"})
+    @JUnitGroup(name = "acceptance")
     public void testConnect() {
         String clusterName = "test-connect";
         oc.newApp("strimzi-connect", map("CLUSTER_NAME", clusterName,
@@ -156,7 +151,7 @@ public class OpenShiftTemplatesIT {
     }
 
     @Test
-    @JUnitGroup(value = {"acceptance"})
+    @JUnitGroup(name = "acceptance")
     public void testS2i() {
         String clusterName = "test-s2i";
         oc.newApp("strimzi-connect-s2i", map("CLUSTER_NAME", clusterName,
@@ -169,7 +164,7 @@ public class OpenShiftTemplatesIT {
     }
 
     @Test
-    @JUnitGroup(value = {"acceptance"})
+    @JUnitGroup(name = "acceptance")
     public void testTopicController() {
         String topicName = "test-topic-cm";
         String mapName = "test-topic-cm-foo";
